- Added Documentation and improved exception handling
