+++ src/sparql/QueryWriter.cpp
@@ -17,14 +17,12 @@
 // along with osm-live-updates.  If not, see <https://www.gnu.org/licenses/>.
 
 #include "sparql/QueryWriter.h"
-#include "sparql/QueryWriterException.h"
 #include "config/Constants.h"
 #include "util/XmlReader.h"
 
 #include <string>
 #include <vector>
-
-//TODO: Check for valid triples
+#include <iostream>
 
 // _________________________________________________________________________________________________
 std::string olu::sparql::QueryWriter::writeInsertQuery(const std::vector<std::string>& triples) {
@@ -49,41 +47,40 @@ std::string olu::sparql::QueryWriter::writeDeleteQuery(const std::string& subjec
 }
 
 // _________________________________________________________________________________________________
-std::string olu::sparql::QueryWriter::getSubjectFor(const std::string &elementTag,
-                                                    const boost::property_tree::ptree &element) {
+std::string
+olu::sparql::QueryWriter::getSubjectFor(const std::string &elementTag,
+                                        const boost::property_tree::ptree &element) {
     std::string identifier;
-    if (elementTag == config::constants::NODE_TAG) {
+    try {
         identifier = olu::util::XmlReader::readAttribute(
                 olu::config::constants::ID_ATTRIBUTE,
                 element);
-        return config::constants::NODE_SUBJECT + ":" + identifier;
+    } catch (std::exception &e) {
+        std::cerr << e.what() << std::endl;
+        std::string msg = "Could not identifier of element: " + util::XmlReader::readTree(element);
+        throw QueryWriterException(msg.c_str());
+    }
+
+    if (elementTag == config::constants::NODE_TAG) {
+        return "osmnode:" + identifier;
     }
 
     if (elementTag == config::constants::WAY_TAG) {
-        identifier = olu::util::XmlReader::readAttribute(
-                olu::config::constants::ID_ATTRIBUTE,
-                element);
         return config::constants::WAY_SUBJECT + ":" + identifier;
     }
 
     if (elementTag == config::constants::RELATION_TAG) {
-        identifier = olu::util::XmlReader::readAttribute(
-                olu::config::constants::ID_ATTRIBUTE,
-                element);
         return config::constants::RELATION_SUBJECT + ":" + identifier;
     }
 
-    throw QueryWriterException(
-            ("Could not determine subject for element: " +
-                    util::XmlReader::readTree(element)).c_str());
+    std::string msg = "Could not determine subject for element: "
+            + util::XmlReader::readTree(element);
+    throw QueryWriterException(msg.c_str());
 }
 
 // _________________________________________________________________________________________________
-std::string olu::sparql::QueryWriter::writeQueryForNodeLocation(const int &nodeId) {
-    std::string query = "SELECT ?o WHERE { "
-             + config::constants::NODE_SUBJECT + ":" + std::to_string(nodeId) + " "
-             + config::constants::LOCATION_AS_WKT_PREDICATE + " "
-             "?o . }";
+std::string olu::sparql::QueryWriter::writeQueryForNodeLocation(const long long &nodeId) {
+    std::string query = "SELECT ?o WHERE { osmnode:" + std::to_string(nodeId) + " geo:hasGeometry/geo:asWKT ?o . }";
     return query;
 }
